# docs/swagger.yaml
definitions:
  model.APIError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: invalid request
        type: string
      requestId:
        example: 7db692fa-c767-468f-af3d-9231b0f88c69
        type: string
      timestamp:
        example: "2025-10-05T13:45:00Z"
        type: string
    type: object
...
  model.ContestSwagger:
    properties:
      awayTeam:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      homeTeam:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        type: string
      quarterResults:
        items:
          $ref: '#/definitions/model.QuarterResult'
        type: array
      squares:
        items:
          $ref: '#/definitions/model.Square'
        type: array
      status:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
      xLabels:
        items:
          type: integer
        type: array
      yLabels:
        items:
          type: integer
        type: array
    type: object
  model.CreateContestRequest:
    properties:
      awayTeam:
        maxLength: 20
        type: string
      homeTeam:
        maxLength: 20
        type: string
      maxSquares:
        maximum: 100
        minimum: 1
        type: integer
      name:
        maxLength: 20
        minLength: 1
        type: string
      owner:
        maxLength: 255
        type: string
      visibility:
        enum:
        - private
        - public
        type: string
    required:
    - maxSquares
    - name
    - owner
    type: object
...
  model.PaginatedContestResponseSwagger:
    properties:
      contests:
        items:
          $ref: '#/definitions/model.ContestSwagger'
        type: array
      hasNext:
        type: boolean
      hasPrevious:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  model.QuarterResult:
    properties:
      awayTeamScore:
        type: integer
      contestId:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      homeTeamScore:
        type: integer
      id:
        type: string
      quarter:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
      winner:
        type: string
      winnerCol:
        type: integer
      winnerName:
        type: string
      winnerRow:
        type: integer
    type: object
  model.QuarterResultRequest:
    properties:
      awayTeamScore:
        maximum: 9999
        minimum: 0
        type: integer
      homeTeamScore:
        maximum: 9999
        minimum: 0
        type: integer
    type: object
  model.Square:
    properties:
      col:
        type: integer
      contestId:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      owner:
        type: string
      ownerName:
        type: string
      row:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
      value:
        type: string
    type: object
...
  model.UpdateContestRequest:
    properties:
      awayTeam:
        maxLength: 20
        type: string
      homeTeam:
        maxLength: 20
        type: string
      visibility:
        enum:
        - private
        - public
        type: string
    type: object
...
  model.UpdateSquareRequest:
    properties:
      owner:
        type: string
      value:
        maxLength: 3
        minLength: 1
        type: string
    required:
    - owner
    - value
    type: object
...
info:
  contact: {}
  description: API for squares.maxstash.io
  title: Squares API
  version: 1.0.0
paths:
...
  /contests:
    put:
      consumes:
      - application/json
      description: Creates a new 10x10 contest
      parameters:
      - description: Contest
        in: body
        name: contest
        required: true
        schema:
          $ref: '#/definitions/model.CreateContestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ContestSwagger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Create a new Contest
      tags:
      - contests
  /contests/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a contest by id. Only the contest owner can delete
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Contest deleted successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Delete contest
      tags:
      - contests
    patch:
      consumes:
      - application/json
      description: Updates the values of a contest
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      - description: Contest update data
        in: body
        name: contest
        required: true
        schema:
          $ref: '#/definitions/model.UpdateContestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ContestSwagger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Update contest
      tags:
      - contests
...
  /contests/{id}/quarter-result:
    post:
      consumes:
      - application/json
      description: Records the score and winner for a specific quarter
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      - description: Quarter result data
        in: body
        name: quarterResult
        required: true
        schema:
          $ref: '#/definitions/model.QuarterResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.QuarterResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Record quarter result
      tags:
      - contests
  /contests/{id}/squares/{squareId}:
    patch:
      consumes:
      - application/json
      description: Updates the value of a specific square in a contest
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      - description: Square ID
        in: path
        name: squareId
        required: true
        type: string
      - description: Square
        in: body
        name: square
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSquareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Square'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Update a single square in a contest
      tags:
      - contests
  /contests/{id}/squares/{squareId}/clear:
    post:
      consumes:
      - application/json
      description: Clears a square's value and owner, making it available for anyone
        to claim
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      - description: Square ID
        in: path
        name: squareId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Square'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Clear square value and owner
      tags:
      - contests
  /contests/{id}/start:
    post:
      consumes:
      - application/json
      description: Starts the contest, transitioning from ACTIVE to Q1 and randomizing
        labels
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ContestSwagger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Start contest
      tags:
      - contests
...
  /contests/owner/{owner}:
    get:
      description: Returns all contests created by a specific owner with pagination
      parameters:
      - description: Owner
        in: path
        name: owner
        required: true
        type: string
      - description: Page number
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: Items per page (max 25)
        in: query
        maximum: 25
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Filter contests by name (case-insensitive)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PaginatedContestResponseSwagger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Get all contests by owner
      tags:
      - contests
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"